move behavioral files to common
