+++ v2/client/src/actions/users.js
@@ -17,21 +17,6 @@ export const fetchUserResults = id => async dispatch => {
   }
 };
 
-export const fetchbehavioralInsight = (role, idOrPIN) => async dispatch => {
-  try {
-    const url = `/api/behavioral-insight/${role}/${idOrPIN}`;
-    const res = await axios.get(url);
-
-    dispatch({
-      type: types.FETCH_BEHAVIORAL_INSIGHT,
-      payload: { data: res.data, role },
-    });
-  } catch (error) {
-    // must check the error status code before
-    history.push('/404err');
-  }
-};
-
 export const fetchLocalLeads = () => async dispatch => {
   try {
     const res = await axios.get('/api/local-leads');
